add video sequence to picture

+++ write.php
@@ -136,6 +136,51 @@ input[type=submit]:hover {
       }
     }
   ?>
+  <h2>Switch Video</h2>
+  <form action="?submit=7" method="post" enctype="multipart/form-data">
+    <div class="row">
+      <h3>Status <?php 
+      if(@$_GET["submit"]==7){
+        $fileOpen = file_get_contents('kas/kas.json');
+        $data = json_decode($fileOpen, TRUE);
+        $data["statusvideo"] = !$data["statusvideo"];
+        $newJsonString = json_encode($data);
+        file_put_contents('kas/kas.json', $newJsonString);
+      }
+      $fileOpen = file_get_contents('kas/kas.json');
+      $data = json_decode($fileOpen, TRUE);
+      if($data["statusvideo"]){
+        echo "On";
+      }
+      else{
+        echo "Off";
+      }
+    ?>
+    </h3>
+    </div>
+    <div class="row">
+      <?php 
+      $fileOpen = file_get_contents('kas/kas.json');
+      $data = json_decode($fileOpen, TRUE);
+      if($data["statusvideo"]){
+        echo "<input type='submit' value='Turn Off'>";
+      }
+      else{
+        echo "<input type='submit' value='Turn On'>";
+      }
+      ?>
+    </div>
+  </form>
+  <?php
+    if(@$_GET["submit"]==7){
+      if($data["statusvideo"] == 1){
+          echo "<h4>Success turn on</h4>";
+      }
+      else{
+        echo "<h4>Success turn off</h4>";
+      }
+    }
+  ?>
   <h2>Jam Tidur</h2>
   <form action="?submit=6" method="post">
     <div class="row">
@@ -255,13 +300,48 @@ input[type=submit]:hover {
       <div class="col-75">
         <input type="file" name="image" id="image" accept="image/gif, image/jpeg, image/png">
       </div>
+    </div>
+    <div class="row">
+      <div class="col-25">
+        <label for="fname">Video</label>
+      </div>
+      <div class="col-75">
+        <input type="file" name="video" id="video" accept="video/mp4">
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-25">
+        <label for="fname">Durasi Gambar</label>
+      </div>
+      <div class="col-75">
+        <?php 
+        $fileOpen = file_get_contents('kas/kas.json');
+        $data = json_decode($fileOpen, TRUE);
+        if(@$_GET["submit"]==3){
+          $fileOpen = file_get_contents('kas/kas.json');
+          $data = json_decode($fileOpen, TRUE);
+          $data["durasigambar"] = $_POST["durasi"];
+          $newJsonString = json_encode($data);
+          file_put_contents('kas/kas.json', $newJsonString);
+          if(!(!file_exists($_FILES['image']['tmp_name']) || !is_uploaded_file($_FILES['image']['tmp_name']))){
+            move_uploaded_file($_FILES["image"]["tmp_name"], __DIR__ . "/kabah.jpg");
+          }
+          if(!(!file_exists($_FILES['video']['tmp_name']) || !is_uploaded_file($_FILES['video']['tmp_name']))){
+            move_uploaded_file($_FILES["video"]["tmp_name"], __DIR__ . "/kabah.mp4");
+          }
+        }
+        if($data["durasigambar"]){
+          echo "<input type='number' name='durasi' value='".$data["durasigambar"]."'>";
+        }
+        ?>
+      </div>
+    </div>
     <div class="row">
       <input type="submit" value="Submit">
     </div>
   </form>
   <?php
     if(@$_GET["submit"]==3){
-      move_uploaded_file($_FILES["image"]["tmp_name"], __DIR__ . "/kabah.jpg");
       echo "<h4>Success uploaded</h4>";
     }
   ?>
